+++ R/fit_model.R
@@ -258,10 +258,23 @@ fit_model_i <- function(data_i,
   # Anomalies should be checked in
   # subsequent steps, not during fitting
   # the model to many datasets.
+  # TODO:
+  # - Update the following lines to
+  #   support using a parameter table.
   if (is.null(model)) {
     model_to_fit <- data_i$model_final
   } else {
     model_to_fit <- model
+  }
+  tmp <- attr(model_to_fit, "ptable")
+  if (!is.null(tmp) &&
+      is.null(data_i$number_of_indicators[[1]]) &&
+      identical(getNamespaceName(environment(fit_function)),
+                c(name = "lavaan"))) {
+    # Do this only for lavaan
+    # TODO:
+    # - Support moderated mediation model with indicators.
+    model_to_fit <- tmp
   }
   # For single-group models,
   # data_i$group_name would be NULL.